[#15] Renaming to Announcer.
