+++ .ruby-gemset
@@ -1,1 +1,1 @@
-ribbon-event_bus
+announcer
