+++ lib/announcer/mixins.rb
@@ -1,0 +1,7 @@
+module Announcer
+  module Mixins
+    autoload(:HasInstance, 'announcer/mixins/has_instance')
+    autoload(:HasConfig, 'announcer/mixins/has_config')
+    autoload(:Serializable, 'announcer/mixins/serializable')
+  end
+end
