+++ lib/announcer/plugins.rb
@@ -1,0 +1,6 @@
+module Announcer
+  module Plugins
+    autoload(:Plugin,        'announcer/plugins/plugin')
+    autoload(:LoggingPlugin, 'announcer/plugins/logging_plugin')
+  end
+end
