+++ lib/announcer/version.rb
@@ -1,0 +1,3 @@
+module Announcer
+  VERSION = "0.5.1"
+end
