--- lib/ribbon/event_bus/version.rb
@@ -1,5 +1,0 @@
-module Ribbon
-  module EventBus
-    VERSION = "0.5.1"
-  end
-end
