--- lib/tasks/event_bus_tasks.rake
@@ -1,4 +1,0 @@
-# desc "Explaining what the task does"
-# task :event_bus do
-#   # Task goes here
-# end
