+++ spec/publishers/async_resque_publisher_spec.rb
@@ -1,7 +1,7 @@
 require 'securerandom'
 require 'redis'
 
-module Ribbon::EventBus
+module Announcer
   module Publishers
     RSpec.describe AsyncResquePublisher do
       before(:all) do
@@ -18,7 +18,7 @@ module Ribbon::EventBus
         ex.run
       }
 
-      let(:instance) { EventBus.instance("resque_test_#{SecureRandom.hex}") }
+      let(:instance) { Announcer.instance("resque_test_#{SecureRandom.hex}") }
       let(:event) { Event.new(:test, instance: instance) }
       let(:subscription) { instance.subscribe_to(:test, priority: 1) { |e| @subscriptions_run = true } }
       let(:resque_mock) { nil }
