+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
-require 'ribbon/event_bus'
+require 'announcer'
 
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
